lib: Support vendor unique commands
When an ATA command is not found in the list of ACS defined commands,
assume that the ATA command is a vendor unique one, non-NCQ and using
the DMA protocol. Such command may define the data transfer size in
bytes or in number of logical sectors. For this, new command flags that
are passed to the ptio_exec_cmd() function and used to generate the
correct ATA PASSTHROUGH SCSI command for executing the ATA command.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>