lib: Fix READ CAPACITY(16) command call
READ CAPACITY(16) fails with some drives because CDB byte 2 that is
set to a nonzero value for the preceding INQUIRY command is
interpreted by the drive as the MSB of the command LBA. Even though the
LBA field in READ CAPACITY(16) is now marked as obsolete and should be
ignored by the drive firmware, some drives do validate it and in this
case the command is bound to fail.

Fix this by zeroing the CDB before initializing it for READ CAPACITY(16)
call.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>